ASCII Arts de las ubicaciones añadidas

+++ scripts/location.sh
@@ -2,6 +2,7 @@
 
 ### Variables ###
 LOCATION=$1
+LOCATION_DIR=$SCRIPTS_DIR/world/$LOCATION
 PROMPT_LOCATION="${GREEN}${PLAYER}@world${NC}:${BLUE}/$LOCATION${NC}$ "
 
 ### Main ###
@@ -9,6 +10,7 @@ function main {
     while true
     do
         clear
+        cat $ASCII_DIR/$LOCATION.ascii | pv -qL 5000
         instructions_location
         if location
         then
@@ -24,7 +26,7 @@ function instructions_location {
     echo -e "${YELLOW}ls${NC}: para listar los archivos con los que puedes hablar."
     echo -e "${YELLOW}cd${NC}: sin argumentos para regresar al mapa mundi."
     echo -e "${YELLOW}cat [ARCHIVO]${NC}: para hablar con dicho archivo."
-    echo -e "${YELLOW}rm [ARCHIVO(s)]${NC}: para ${RED}eliminar${NC} uno o varios archivos."
+    echo -e "${YELLOW}rm [ARCHIVO(s)]${NC}: para ${RED}eliminar${NC} uno o varios archivos. Usa * para eliminar todos a la vez."
     echo -e "${YELLOW}help${NC}: para mostrar las instrucciones de nuevo."
     echo ""
 }
@@ -59,7 +61,13 @@ function location {
                     FILE="$(echo "${COMMAND}" | cut -d' ' -f2)"
                     if [[ -f $FILE ]]
                     then
-                        source $SCRIPTS_DIR/world/$LOCATION/$FILE
+                        source $LOCATION_DIR/$FILE
+                        if grep -Fxq "${FILE}" $TEMP/$LOCATION
+                            then
+                                :
+                            else
+                                echo $FILE >> $TEMP/$LOCATION
+                        fi
                         return 1
                     else
                         echo "cat: $FILE: No existe el fichero o el directorio"
@@ -85,31 +93,36 @@ function location {
                 fi
             ;;
             "rm"*)
-                ARGS=$(wc -w <<< "$COMMAND")
-                if [ $ARGS -eq 1 ]
+                if [[ $(ls $LOCATION_DIR | sort) = $(sort $TEMP/$LOCATION 2> /dev/null) ]]
                 then
-                    echo "rm: falta un operando"
-                elif [ $ARGS -gt 1 ]
-                then
-                    FILES="$(echo "${COMMAND}" | cut -d' ' -f2-)"
-                    IFS=" " read -r -a FILES_ARRAY <<< "${FILES}"
-                    VALUE='^\*$'
-                    if [[ "${FILES_ARRAY[@]}" =~ $VALUE ]]
+                    ARGS=$(wc -w <<< "$COMMAND")
+                    if [ $ARGS -eq 1 ]
                     then
-                        echo -e "${RED}Eliminando${NC} todos los archivos"
-                        rm *
-                    else
-                        for FILE in "${FILES_ARRAY[@]}"
-                        do
-                            if [[ -f $FILE ]]
-                            then
-                                rm $FILE
-                                echo "El archivo ${FILE} ha sido eliminado"
-                            else
-                                echo "rm: ${FILE}: No existe el fichero o el directorio"
-                            fi
-                        done
+                        echo "rm: falta un operando"
+                    elif [ $ARGS -gt 1 ]
+                    then
+                        FILES="$(echo "${COMMAND}" | cut -d' ' -f2-)"
+                        IFS=" " read -r -a FILES_ARRAY <<< "${FILES}"
+                        VALUE='^\*$'
+                        if [[ "${FILES_ARRAY[@]}" =~ $VALUE ]]
+                        then
+                            echo -e "${RED}Eliminando${NC} todos los archivos"
+                            rm *
+                        else
+                            for FILE in "${FILES_ARRAY[@]}"
+                            do
+                                if [[ -f $FILE ]]
+                                then
+                                    rm $FILE
+                                    echo "El archivo ${FILE} ha sido eliminado"
+                                else
+                                    echo "rm: ${FILE}: No existe el fichero o el directorio"
+                                fi
+                            done
+                        fi
                     fi
+                else
+                    echo "rm: Primero debes hablar con todos los lugareños para poder usar este comando"
                 fi
             ;;
             "help")
